Move MBeanQuery out of package entity
