+++ java.d/src/test/java/org/firehol/netdata/module/jmx/MBeanServerCollectorTest.java
@@ -44,7 +44,6 @@ import org.firehol.netdata.model.DimensionAlgorithm;
 import org.firehol.netdata.module.jmx.configuration.JmxChartConfiguration;
 import org.firehol.netdata.module.jmx.configuration.JmxDimensionConfiguration;
 import org.firehol.netdata.module.jmx.configuration.JmxServerConfiguration;
-import org.firehol.netdata.module.jmx.entity.MBeanQuery;
 import org.firehol.netdata.module.jmx.exception.JmxMBeanServerQueryException;
 import org.firehol.netdata.test.utils.ReflectionUtils;
 import org.firehol.netdata.test.utils.TestObjectBuilder;
